move some shared code to curve-lib
