+++ apps/lend/src/store/createUserSlice.ts
@@ -4,7 +4,7 @@ import type { State } from '@/store/useStore'
 import cloneDeep from 'lodash/cloneDeep'
 
 import apiLending, { helpers } from '@/lib/apiLending'
-import { log } from '@/utils/helpers'
+import { log } from '@/shared/curve-lib'
 
 type StateKey = keyof typeof DEFAULT_STATE
