+++ apps/main/src/entities/chain/model/chain-validation.ts
@@ -1,5 +1,5 @@
 import type { ChainQueryParams } from '@/entities/chain/types'
-import { createValidationSuite } from '@/shared/validation'
+import { createValidationSuite } from 'curve-lib/src/shared/validation'
 import { enforce, group, test } from 'vest'
 
 export const chainValidationGroup = ({ chainId }: ChainQueryParams) =>
