+++ apps/main/src/entities/gauge/api/gauge-query.ts
@@ -13,7 +13,7 @@ import { assertGaugeValidity } from '@/entities/gauge/lib'
 import { GaugeQueryKeyType, type PoolMethodResult } from '@/entities/gauge/types'
 import { BD } from '@/shared/curve-lib'
 import useStore from '@/store/useStore'
-import { logQuery } from '@/utils'
+import { logQuery } from '@/shared/curve-lib'
 import { QueryFunction } from '@tanstack/react-query'
 import { zeroAddress, type Address } from 'viem'
