+++ apps/main/src/entities/gauge/lib/reward-actions.ts
@@ -20,7 +20,7 @@ import type {
   PoolMethodResult,
 } from '@/entities/gauge/types'
 import useTokensMapper from '@/hooks/useTokensMapper'
-import { queryClient } from '@/shared/api/query-client'
+import { queryClient } from '@/shared/curve-lib'
 import useStore from '@/store/useStore'
 import { t } from '@lingui/macro'
 import { useIsMutating, useMutation, UseMutationResult } from '@tanstack/react-query'
