+++ apps/main/src/entities/pool/model/pool-validation.ts
@@ -1,6 +1,6 @@
 import { chainValidationGroup } from '@/entities/chain'
 import type { PoolQueryParams } from '@/entities/pool/types'
-import { createValidationSuite } from '@/shared/validation'
+import { createValidationSuite } from '../../../../../../packages/curve-lib/src/shared/validation'
 import { enforce, group, test } from 'vest'
 
 export const poolValidationGroup = ({ chainId, poolId }: PoolQueryParams) =>
