+++ apps/main/src/features/add-gauge-reward-token/model/validation-suite.ts
@@ -1,6 +1,6 @@
 import { gaugeAddRewardTokenValidationGroup } from '@/entities/gauge'
 import type { AddRewardFormValues } from '@/features/add-gauge-reward-token/types'
-import { createValidationSuite } from '@/shared/validation'
+import { createValidationSuite } from '../../../../../../packages/curve-lib/src/shared/validation'
 
 export const addGaugeRewardTokenValidationSuite = createValidationSuite<AddRewardFormValues>(
   gaugeAddRewardTokenValidationGroup
