+++ apps/main/src/features/add-gauge-reward-token/ui/AddRewardToken.tsx
@@ -10,8 +10,8 @@ import TxInfoBar from '@/ui/TxInfoBar'
 
 import { useSignerAddress } from '@/entities/signer'
 import { DistributorInput, EstimatedGasInfo, FormActions, TokenSelector } from '@/features/add-gauge-reward-token/ui'
-import { formDefaultOptions } from '@/shared/model/form'
-import { FormErrorsDisplay } from 'shared/ui/forms'
+import { formDefaultOptions } from '@/shared/curve-lib'
+import { FormErrorsDisplay } from '@/ui/FormErrorsDisplay'
 import { FlexContainer, FormContainer, FormFieldsContainer } from '@/shared/ui/styled-containers'
 import { vestResolver } from '@hookform/resolvers/vest'
 import { zeroAddress } from 'viem'
