+++ apps/main/src/store/createPoolsSlice.ts
@@ -27,7 +27,8 @@ import isNaN from 'lodash/isNaN'
 import pick from 'lodash/pick'
 
 import { INVALID_ADDRESS } from '@/constants'
-import { fulfilledValue, getChainPoolIdActiveKey, getCurvefiUrl, log } from '@/utils'
+import { fulfilledValue, getChainPoolIdActiveKey, getCurvefiUrl } from '@/utils'
+import { log } from '@/shared/curve-lib'
 import { convertToLocaleTimestamp } from '@/ui/Chart/utils'
 import curvejsApi from '@/lib/curvejs'
 import networks from '@/networks'
