+++ apps/main/src/widgets/manage-gauge/ui/ManageGauge.tsx
@@ -2,7 +2,7 @@ import { useGaugeManager, useGaugeRewardsDistributors } from '@/entities/gauge'
 import { useSignerAddress } from '@/entities/signer'
 import AddRewardToken from '@/features/add-gauge-reward-token'
 import DepositReward from '@/features/deposit-gauge-reward'
-import TabSlider, { Tab } from 'shared/ui/tab-slider/TabSlider'
+import TabSlider, { Tab } from '@/ui/TabSlider/TabSlider'
 import type { TabValue } from '@/widgets/manage-gauge/types'
 import { useCallback, useMemo, useState } from 'react'
 import { isAddressEqual, type Address } from 'viem'
