+++ packages/curve-lib/src/index.ts
@@ -1,4 +1,4 @@
 export * from './bigNumber'
 export * from './web3'
 export * from './network'
-export * from './shared/big-decimal'
+export * from './shared'
