+++ packages/curve-lib/src/shared/api/query-provider.tsx
