+++ packages/curve-lib/src/shared/lib/queries/types.ts
