+++ packages/curve-lib/src/shared/types/api.ts
