+++ packages/curve-lib/src/shared/types/index.ts
@@ -1,0 +1,2 @@
+export * from './api'
+export * from './nested'
