+++ packages/curve-lib/src/shared/types/nested.ts
