+++ packages/curve-lib/src/shared/validation/basic-validation.ts
