+++ packages/curve-lib/src/shared/validation/index.ts
