+++ packages/curve-lib/src/shared/validation/types.ts
