+++ packages/curve-lib/src/shared/validation/validation.d.ts
