+++ packages/ui/src/index.tsx
@@ -18,6 +18,7 @@ export * from './DetailInfo'
 export * from './Dialog'
 export * from './DialogComboBox'
 export * from './DividerHorizontal'
+export * from './FormErrorsDisplay'
 export * from './HeaderStats'
 export * from './Icon'
 export * from './IconButton'
@@ -46,6 +47,7 @@ export * from './Stepper'
 export * from './Switch'
 export * from './Tab'
 export * from './TabSlide'
+export * from './TabSlider'
 export * from './Table'
 export * from './TableButtonFilters'
 export * from './TableButtonFiltersMobile'
